+++ src/messages/nl/audit.php
@@ -60,6 +60,7 @@ return [
     'Mails' => 'Emails',
     'Memory' => 'Geheugen',
     'Message' => 'Bericht',
+    'Model ID' => 'Model ID',
     'No' => 'Neen',
     'Origin' => 'Afkomst',
     'Profiling' => '',