+++ tests/codeception/unit/AuditEntryTest.php
@@ -4,6 +4,7 @@ namespace tests\codeception\unit;
 
 use bedezign\yii2\audit\models\AuditData;
 use bedezign\yii2\audit\models\AuditEntry;
+use bedezign\yii2\audit\tests\UnitTester;
 use Codeception\Specify;
 
 /**
@@ -20,55 +21,70 @@ class AuditEntryTest extends AuditTestCase
 
     public function testEnabledBatchSaveOptionIsRespected()
     {
-        $module = $this->module();
-        $module->batchSave = true;
+        $this->module()->batchSave = true;
 
-        $mock = $this->getMock(AuditEntry::className());
+        $mock = $this->getMockBuilder(AuditEntry::className())->getMock();
         $mock->expects($this->once())->method('addBatchData');
         $mock->expects($this->never())->method('addData');
-        $this->useEntry($mock);
 
+        $this->finalizeAudit();
+        $this->useEntry($mock);
         $this->finalizeAudit(false);
     }
 
     public function testBatchSaveOptionIsRespected()
     {
-        $module = $this->module();
-        $module->batchSave = false;
+        $this->module()->batchSave = false;
 
-        $mock = $this->getMock(AuditEntry::className());
+        $mock = $this->getMockBuilder(AuditEntry::className())->getMock();
         $mock->expects($this->never())->method('addBatchData');
         $mock->expects($this->atLeastOnce())->method('addData');
-        $this->useEntry($mock);
 
+        $this->finalizeAudit();
+        $this->module()->batchSave = false;
+        $this->useEntry($mock);
         $this->finalizeAudit(false);
     }
 
     public function testThatAddDataWorks()
     {
-        $oldId = $this->tester->fetchTheLastModelPk(AuditData::className());
-        $module = $this->module();
-        $module->batchSave = false;
-        $entry = $this->entry();
-
+        $this->module()->batchSave = false;
+        $this->entry();
         $this->finalizeAudit();
+        $this->module()->batchSave = false;
 
-        $newId = $this->tester->fetchTheLastModelPk(AuditData::className());
-        $this->assertGreaterThan($oldId, $newId);
+        $oldEntry = $this->entry();
+        $this->finalizeAudit();
+        $this->module()->batchSave = false;
         $this->tester->seeRecord(AuditData::className(), [
-            'entry_id' => $entry->id,
+            'entry_id' => $oldEntry->id,
             'type' => 'audit/request',
         ]);
+
+        $newEntry = $this->entry();
+        $this->finalizeAudit();
+        $this->module()->batchSave = false;
+        $this->tester->seeRecord(AuditData::className(), [
+            'entry_id' => $newEntry->id,
+            'type' => 'audit/request',
+        ]);
+
+        $this->assertGreaterThan($oldEntry->id, $newEntry->id);
     }
 
     public function testBatchSaveOptionIsWorksUncompressed()
     {
-        $module = $this->module();
-        $module->batchSave = false;
-        $module->compressData = false;
+        $this->module()->batchSave = true;
+        $this->module()->compressData = false;
+        $this->entry();
+        $this->finalizeAudit();
+        $this->module()->batchSave = true;
+        $this->module()->compressData = false;
 
         $entry = $this->entry();
         $this->finalizeAudit();
+        $this->module()->batchSave = true;
+        $this->module()->compressData = false;
 
         $this->tester->seeRecord(AuditData::className(), [
             'entry_id' => $entry->id,
@@ -78,11 +94,17 @@ class AuditEntryTest extends AuditTestCase
 
     public function testThatAddDataWorksUncompressed()
     {
-        $module = $this->module();
-        $module->compressData = false;
+        $this->module()->batchSave = false;
+        $this->module()->compressData = false;
+        $this->entry();
+        $this->finalizeAudit();
+        $this->module()->batchSave = false;
+        $this->module()->compressData = false;
 
         $entry = $this->entry();
         $this->finalizeAudit();
+        $this->module()->batchSave = false;
+        $this->module()->compressData = false;
 
         $this->tester->seeRecord(AuditData::className(), [
             'entry_id' => $entry->id,