+++ tests/codeception/unit/AuditTrailBehaviorWebTest.php
@@ -30,16 +30,16 @@ class AuditTrailBehaviorWebTest extends AuditTestCase
      */
     public function testCreatePost()
     {
-        $oldEntryId = $this->tester->fetchTheLastModelPk(AuditEntry::className());
+        $this->entry();
+        $this->finalizeAudit();
+
         $post = new Post();
         $post->title = 'New post title';
         $post->body = 'New post body';
         $this->assertTrue($post->save());
 
+        $entry = $this->entry();
         $this->finalizeAudit();
-
-        $newEntryId = $this->tester->fetchTheLastModelPk(AuditEntry::className());
-        $this->assertNotEquals($oldEntryId, $newEntryId, 'I expected that a new entry was added');
 
         $this->tester->seeRecord(Post::className(), [
             'id' => $post->id,
@@ -47,11 +47,11 @@ class AuditTrailBehaviorWebTest extends AuditTestCase
             'body' => 'New post body',
         ]);
         $this->tester->seeRecord(AuditEntry::className(), [
-            'id' => $newEntryId,
+            'id' => $entry->id,
             'request_method' => 'GET',
         ]);
         $this->tester->seeRecord(AuditTrail::className(), [
-            'entry_id' => $newEntryId,
+            'entry_id' => $entry->id,
             'action' => 'CREATE',
             'model' => Post::className(),
             'model_id' => $post->id,
@@ -60,7 +60,7 @@ class AuditTrailBehaviorWebTest extends AuditTestCase
             'new_value' => $post->id,
         ]);
         $this->tester->seeRecord(AuditTrail::className(), [
-            'entry_id' => $newEntryId,
+            'entry_id' => $entry->id,
             'action' => 'CREATE',
             'model' => Post::className(),
             'model_id' => $post->id,
@@ -69,7 +69,7 @@ class AuditTrailBehaviorWebTest extends AuditTestCase
             'new_value' => 'New post title',
         ]);
         $this->tester->seeRecord(AuditTrail::className(), [
-            'entry_id' => $newEntryId,
+            'entry_id' => $entry->id,
             'action' => 'CREATE',
             'model' => Post::className(),
             'model_id' => $post->id,
@@ -84,16 +84,20 @@ class AuditTrailBehaviorWebTest extends AuditTestCase
      */
     public function testUpdatePost()
     {
-        $oldEntryId = $this->tester->fetchTheLastModelPk(AuditEntry::className());
+        $this->entry();
+        $this->finalizeAudit();
+
+        $oldEntryId = $this->getLastPk(AuditEntry::className());
 
         $post = Post::findOne(1);
         $post->title = 'Updated post title';
         $post->body = 'Updated post body';
         $this->assertTrue($post->save());
 
+        $this->entry();
         $this->finalizeAudit();
 
-        $newEntryId = $this->tester->fetchTheLastModelPk(AuditEntry::className());
+        $newEntryId = $this->getLastPk(AuditEntry::className());
         $this->assertNotEquals($oldEntryId, $newEntryId, 'I expected that a new entry was added');
 
         $this->tester->seeRecord(Post::className(), [
@@ -130,13 +134,17 @@ class AuditTrailBehaviorWebTest extends AuditTestCase
      */
     public function testDeletePost()
     {
-        $oldEntryId = $this->tester->fetchTheLastModelPk(AuditEntry::className());
+        $this->entry();
+        $this->finalizeAudit();
+
+        $oldEntryId = $this->getLastPk(AuditEntry::className());
         $post = Post::findOne(1);
         $this->assertSame($post->delete(), 1);
 
+        $this->entry();
         $this->finalizeAudit();
 
-        $newEntryId = $this->tester->fetchTheLastModelPk(AuditEntry::className());
+        $newEntryId = $this->getLastPk(AuditEntry::className());
         $this->assertNotEquals($oldEntryId, $newEntryId, 'I expected that a new entry was added');
 
         $this->tester->dontSeeRecord(Post::className(), [